requirements.txt
selenium
requests
pytest
allure-pytest
webdriver-manager
flake8